FileRotator:fix NPE due to The File.list() method may be return null. Bug: 337070728 Change-Id: I63707b4f9fe3bcd705e98ac2e69a5101d092de76 Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

+++ core/java/com/android/internal/util/FileRotator.java
@@ -144,7 +144,15 @@ public class FileRotator {
      */
     public void deleteAll() {
         final FileInfo info = new FileInfo(mPrefix);
-        for (String name : mBasePath.list()) {
+        String[] baseFiles = mBasePath.list();
+        if (baseFiles == null) {
+            // no file in the path and create one starting now
+            info.startMillis = currentTimeMillis;
+            info.endMillis = Long.MAX_VALUE;
+            return info.build();
+        }
+
+        for (String name : baseFiles) {
             if (info.parse(name)) {
                 // delete each file that matches parser
                 new File(mBasePath, name).delete();
